% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QVM.R
\name{QVM}
\alias{QVM}
\title{Questionnaires Validation Module}
\usage{
QVM()
}
\value{
QVM is an interface
}
\description{
This package does a multivariate analysis for questionnaire validation with Likert-type scale variables.

You can learn more about this package at:
http://www.uv.mx/personal/nehuerta/qvm/
}
\details{
Introduce a data set of categorical information and classifed as numerical data.
Choose a data set (delimited csv for comma).
This funtion takes a data set and applies questionnaire validatioin techniques to return parameter estimates.
}
\examples{
\dontrun{
##Call to library
library(QVM)
##Call the package
QVM()
}
}
\author{
Nery Sofia Huerta-Pacheco, Purificacion Vicente-Galindo, Mercedes Sanchez Barba,  Maria Purificacion Galindo Villardon
}
\references{
Allen, F., & Locker, D. (2002). A Modified Short Version of the Oral Health Impact
Profile for Assessing Health-Related Quality of Life in Edentulous Adults. The
International Journal of Prosthodontics, 15(5), 446-450.

Baker, F. (2004). Item Response Theory: Parameter Estimation Techniques,
Second Edition. CRC Press.

Cronbach, L. (1951). Coefficient alpha and the internal structure of tests.
Psychomerika, 16, 297–334.

Gil Pascual, J. A. (2015). Metodología cuantitativa en educación. UNED

Juniper, E. F., Guyatt, G. H., Streiner, D. L., & King, D. R. (1997). Clinical impact
versus factor analysis for quality of life questionnaire construction. Journal of
ClinicalEpidemiology, 233-238.

Likert, R. (1932). A Technique for the Measurement of Attitudes. Archives of
Psychology, 140, 1–55.

Martínez Arias, M. R., Hernández Lloreda, M. V., & Hernández Lloreda, M. J. (2014).
Psicometría. Madrid: Alianza Editorial.

Ponsoda, V., Revuelta, J. & Abad, F. J. (2006). Modelos politómicos de respuesta al
ítem. Madrid: La Muralla.

Rizopoulos, D. (2006). ltm: An R package for Latent Variable Modelling and Item
Response Theory Analyses, Journal of Statistical Software, 17 (5), 1-25.
URL http://www.jstatsoft.org/v17/i05/

Stevens, S. S. (1946). On the Theory of Scales of Measurement. Science, 677-680.
}

